Move prelude Namespace into BuiltinRegistry
